Make interface less String-y
Add FieldName trait for the names of fields
Add FieldNames associated type to Entry to indicate what type the field
names are.

The idea is that each entry will define an enum with the field names as
variants. This will ensure that you cannot use an invalid field name to
retrive a field's value, for example.

This also adds a few methods to the Table trait to make finding things
easier